+++ Tests/MapboxNavigationTests/VanishingRouteLineTests.swift
@@ -6,6 +6,7 @@ import MapboxMaps
 @testable import MapboxCoreNavigation
 
 class VanishingRouteLineTests: TestCase {
+    let accuracyThreshold = 1e-5
     var navigationMapView: NavigationMapView!
     var window = UIWindow()
     
@@ -23,10 +24,6 @@ class VanishingRouteLineTests: TestCase {
     
     func getRoute() -> Route {
         let routeData = Fixture.JSONFromFileNamed(name: "short_route")
-        let routeOptions = NavigationRouteOptions(coordinates: [
-            CLLocationCoordinate2DMake(-122.5237429, 37.975393),
-            CLLocationCoordinate2DMake(-122.5231413, 37.9750695)
-        ])
         let decoder = JSONDecoder()
         decoder.userInfo[.options] = routeOptions
         var testRoute: Route?
@@ -56,8 +53,25 @@ class VanishingRouteLineTests: TestCase {
         return route
     }
     
-    func getRouteProgress() -> RouteProgress {
-        let route = getRoute()
+    func getStraightLineRoute() -> Route {
+        let routeData = Fixture.JSONFromFileNamed(name: "route-with-straight-line")
+        let routeOptions = NavigationRouteOptions(coordinates: [
+            CLLocationCoordinate2DMake(55.026291, 37.975393),
+            CLLocationCoordinate2DMake(55.026720, 30.798226)
+        ])
+        let decoder = JSONDecoder()
+        decoder.userInfo[.options] = routeOptions
+        var testRoute: Route?
+        XCTAssertNoThrow(testRoute = try decoder.decode(Route.self, from: routeData))
+        guard let route = testRoute else {
+            preconditionFailure("Route is invalid.")
+        }
+        
+        return route
+    }
+    
+    func getRouteProgress(_ route: Route? = nil) -> RouteProgress {
+        let route = route ?? getRoute()
         let routeProgress = RouteProgress(route: route, options: routeOptions, legIndex: 0, spokenInstructionIndex: 0)
         routeProgress.currentLegProgress = RouteLegProgress(leg: route.legs[0], stepIndex: 2, spokenInstructionIndex: 0)
         routeProgress.currentLegProgress.currentStepProgress = RouteStepProgress(step: route.legs[0].steps[2], spokenInstructionIndex: 0)
@@ -88,6 +102,14 @@ class VanishingRouteLineTests: TestCase {
         return routeProgress
     }
     
+    func updateVanishingRouteLine(route: Route, routeProgress: RouteProgress, coordinate: CLLocationCoordinate2D) {
+        navigationMapView.routeLineTracksTraversal = true
+        navigationMapView.show([route])
+        routeProgress.updateDistanceTraveled(with: .init(coordinate: coordinate))
+        navigationMapView.updateUpcomingRoutePointIndex(routeProgress: routeProgress)
+        navigationMapView.updateFractionTraveled(coordinate: coordinate)
+    }
+    
     func lineGradientToString(lineGradient: Value<StyleColor>?) -> String {
         guard let halfStringFromLineGradient = lineGradient.debugDescription.components(separatedBy: "(").last,
               let stringFromLineGradient = halfStringFromLineGradient.components(separatedBy: ")").first else {
@@ -107,10 +129,10 @@ class VanishingRouteLineTests: TestCase {
         XCTAssertEqual(navigationMapView.mapView.cameraState.zoom, zoomeLevel, "Zoom levels should be equal.")
     }
     
-    func testParseRoutePoints() {
+    func testInitPrimaryRoutePoints() {
         // https://github.com/mapbox/mapbox-navigation-android/blob/0ca183f7cb7bec930521ea9bcd59d0e8e2bef165/libnavui-maps/src/test/java/com/mapbox/navigation/ui/maps/internal/route/line/MapboxRouteLineUtilsTest.kt#L1798
-        let route = getMultilegRoute()
-        let routePoints = navigationMapView.parseRoutePoints(route: route)
+        var route = getMultilegRoute()
+        var routePoints = navigationMapView.parseRoutePoints(route: route)
         
         // Because mapbox-directions-swift parses the route with one more duplicate coordinate in the last step of each route leg.
         // The two leg route has two more coordinates compared with Android.
@@ -118,8 +140,25 @@ class VanishingRouteLineTests: TestCase {
         XCTAssertEqual(routePoints.nestedList.flatMap{$0}.count, 15)
         XCTAssertEqual(routePoints.flatList[1].latitude, routePoints.flatList[2].latitude)
         XCTAssertEqual(routePoints.flatList[1].longitude, routePoints.flatList[2].longitude)
-        XCTAssertEqual(routePoints.flatList[128].latitude, routePoints.flatList[129].latitude, accuracy: 0.000001)
-        XCTAssertEqual(routePoints.flatList[128].longitude, routePoints.flatList[129].longitude, accuracy: 0.000001)
+        XCTAssertEqual(routePoints.flatList[128].latitude, routePoints.flatList[129].latitude, accuracy: accuracyThreshold)
+        XCTAssertEqual(routePoints.flatList[128].longitude, routePoints.flatList[129].longitude, accuracy: accuracyThreshold)
+        
+        route = getStraightLineRoute()
+        routePoints = navigationMapView.parseRoutePoints(route: route)
+        XCTAssertEqual(routePoints.flatList.count, 6)
+        // This route has 3 steps.
+        XCTAssertEqual(routePoints.nestedList.flatMap{$0}.count, 3)
+        
+        guard let routeLineGranularDistances = navigationMapView.calculateGranularDistances(routePoints.flatList) else {
+            XCTFail("Failed to calculate granular distances.")
+            return
+        }
+        let distanceArray = routeLineGranularDistances.distanceArray
+        XCTAssertEqual(distanceArray.count, routePoints.flatList.count)
+        XCTAssertEqual(distanceArray.first?.distanceRemaining, routeLineGranularDistances.distance)
+        XCTAssertEqual(distanceArray.first?.point, routePoints.flatList.first)
+        XCTAssertEqual(distanceArray.last?.distanceRemaining, 0.0)
+        XCTAssertEqual(distanceArray.last?.point, routePoints.flatList.last)
     }
     
     func testUpdateUpcomingRoutePointIndex() {
@@ -127,11 +166,9 @@ class VanishingRouteLineTests: TestCase {
         let route = getRoute()
         
         navigationMapView.initPrimaryRoutePoints(route: route)
-        navigationMapView.routeLineGranularDistances = nil
         XCTAssertEqual(navigationMapView.fractionTraveled, 0.0)
         
         let routeProgress = getRouteProgress()
-        
         navigationMapView.updateUpcomingRoutePointIndex(routeProgress: routeProgress)
         
         // Because mapbox-directions-swift parses the route with one more duplicate coordinate in the last step of each route leg.
@@ -150,33 +187,52 @@ class VanishingRouteLineTests: TestCase {
     
     func testUpdateFractionTraveled() {
         // https://github.com/mapbox/mapbox-navigation-android/blob/0ca183f7cb7bec930521ea9bcd59d0e8e2bef165/libnavui-maps/src/test/java/com/mapbox/navigation/ui/maps/route/line/api/MapboxRouteLineApiTest.kt#L636
-        let routeProgress = getRouteProgress()
-        let route = routeProgress.route
-        
-        let coordinate = route.shape!.coordinates[1]
-        navigationMapView.routeLineTracksTraversal = true
-        navigationMapView.show([route])
-        navigationMapView.updateUpcomingRoutePointIndex(routeProgress: routeProgress)
-        navigationMapView.updateFractionTraveled(coordinate: coordinate)
+        var routeProgress = getRouteProgress()
+        var route = routeProgress.route
+        var coordinate = route.shape!.coordinates[1]
+        updateVanishingRouteLine(route: route, routeProgress: routeProgress, coordinate: coordinate)
         
         // When `routeLineTracksTraversal` enabled, the `fractionTraveled` is expected to be updated after
         // the upcoming route point index update and a location update.
-        let expectedFractionTraveled = 0.3240769449298392
-        XCTAssertEqual(navigationMapView.fractionTraveled, expectedFractionTraveled, accuracy: 0.0000000001)
+        var expectedFractionTraveled = 0.3240769449298392
+        XCTAssertEqual(navigationMapView.fractionTraveled, expectedFractionTraveled, accuracy: accuracyThreshold)
+        
+        route = getStraightLineRoute()
+        routeProgress = RouteProgress(route: route, options: routeOptions, legIndex: 0, spokenInstructionIndex: 0)
+        routeProgress.currentLegProgress = RouteLegProgress(leg: route.legs[0], stepIndex: 0, spokenInstructionIndex: 0)
+        routeProgress.currentLegProgress.currentStepProgress = RouteStepProgress(step: route.legs[0].steps[0], spokenInstructionIndex: 0)
+        routeProgress.currentLegProgress.currentStepProgress.distanceTraveled = 43
+        
+        coordinate = CLLocationCoordinate2D(latitude: 55.02624655478515, longitude: 30.79885117582073)
+        updateVanishingRouteLine(route: route, routeProgress: routeProgress, coordinate: coordinate)
+        
+        let expectedUpcomingPointIndex = 2
+        // When the user location is approaching to the last coordinate of the step, set the upcoming point to the start of the next step.
+        XCTAssertEqual(navigationMapView.routeRemainingDistancesIndex, expectedUpcomingPointIndex)
+        
+        guard let firstCoordinate = route.shape?.coordinates.first,
+              let totalDistance = route.shape?.distance() else {
+            XCTFail("Failed to update vanishing route line.")
+            return
+        }
+        
+        let traveledDistance = firstCoordinate.distance(to: coordinate)
+        let fractionTraveledFromStart = traveledDistance / totalDistance
+        expectedFractionTraveled = 0.3272757239007916
+        // Compare the Haversine calculated result from start coodinate with the project distance calculated result.
+        XCTAssertEqual(fractionTraveledFromStart, expectedFractionTraveled, accuracy: accuracyThreshold)
+        XCTAssertEqual(navigationMapView.fractionTraveled, expectedFractionTraveled, accuracy: accuracyThreshold)
     }
     
     func testEmptyRouteWithValidRouteProgress() {
-        let routeProgress = getRouteProgress()
         let route = getEmptyRoute()
+        let routeProgress = getRouteProgress(route)
         
         let coordinate = CLLocationCoordinate2DMake(-122.5237429, 37.975393)
-        navigationMapView.routeLineTracksTraversal = true
-        navigationMapView.show([route])
-        navigationMapView.updateUpcomingRoutePointIndex(routeProgress: routeProgress)
-        navigationMapView.updateFractionTraveled(coordinate: coordinate)
+        updateVanishingRouteLine(route: route, routeProgress: routeProgress, coordinate: coordinate)
         
-        // Route without coordinates inside its steps would lead to invalid routeRemainingDistancesIndex.
-        XCTAssertEqual(navigationMapView.routeRemainingDistancesIndex, -1)
+        // Route without coordinates inside its steps would lead to 0 routeRemainingDistancesIndex.
+        XCTAssertEqual(navigationMapView.routeRemainingDistancesIndex, 0)
         XCTAssertEqual(navigationMapView.fractionTraveled, 0.0, accuracy: 0)
     }
     
@@ -185,10 +241,7 @@ class VanishingRouteLineTests: TestCase {
         let routeProgress = getUnstartedRouteProgress(route: route)
         
         let coordinate = CLLocationCoordinate2DMake(-122.5237429, 37.975393)
-        navigationMapView.routeLineTracksTraversal = true
-        navigationMapView.show([route])
-        navigationMapView.updateUpcomingRoutePointIndex(routeProgress: routeProgress)
-        navigationMapView.updateFractionTraveled(coordinate: coordinate)
+        updateVanishingRouteLine(route: route, routeProgress: routeProgress, coordinate: coordinate)
         
         XCTAssert(navigationMapView.routeRemainingDistancesIndex! >= 0, "Non-empty route should have valid routeRemainingDistancesIndex.")
         XCTAssertEqual(navigationMapView.fractionTraveled, 0.0, accuracy: 0)
@@ -233,7 +286,7 @@ class VanishingRouteLineTests: TestCase {
         navigationMapView.travelAlongRouteLine(to: coordinate)
         let expectedFractionTraveled = 0.3240769449298392
         let actualFractionTraveled = navigationMapView.fractionTraveled
-        XCTAssertEqual(actualFractionTraveled, expectedFractionTraveled, accuracy: 0.0000000001, "Failed to update route line when routeLineTracksTraversal enabled.")
+        XCTAssertEqual(actualFractionTraveled, expectedFractionTraveled, accuracy: accuracyThreshold, "Failed to update route line when routeLineTracksTraversal enabled.")
         
         let layerIdentifier = route.identifier(.route(isMainRoute: true))
         do {